requirements.txt
PyYAML>=6.0
Markdown>=3.5
Jinja2>=3.1
playwright>=1.44
Flask>=3.0
Pillow>=10.0
pytesseract>=0.3.10
rapidocr-onnxruntime>=1.4.4
requests>=2.31
